f64 -> f32

+++ src/camera.rs
@@ -5,12 +5,12 @@ pub struct SimpleCamera {}
 
 #[allow(non_upper_case_globals)]
 impl SimpleCamera {
-  const aspect_ratio: f64 = 16.0 / 9.0;
-  const viewport_height: f64 = 2.0;
-  const viewport_width: f64 = Self::aspect_ratio * Self::viewport_height;
-  const focal_length: f64 = 1.0;
+  const aspect_ratio: f32 = 16.0 / 9.0;
+  const viewport_height: f32 = 2.0;
+  const viewport_width: f32 = Self::aspect_ratio * Self::viewport_height;
+  const focal_length: f32 = 1.0;
 
-  pub fn ray(u: f64, v: f64) -> Ray {
+  pub fn ray(u: f32, v: f32) -> Ray {
     let origin = Vec3::zero();
     let horizontal = vec(Self::viewport_width, 0.0, 0.0);
     let vertical = vec(0.0, Self::viewport_height, 0.0);
